+++ src/main/java/dukepro/handlers/Interact.java
@@ -64,7 +64,7 @@ public class Interact {
         } else if (word.startsWith("delExpense")) {
             return expenseManager.delete(Decoder.deleteExpense(word, expenseManager.numStored()));
         } else if (word.startsWith("totalSpent")) {
-            int total = expenseManager.operateOnList(arr -> ExpenseCalculator.sumArrayList(arr));
+            double total = expenseManager.operateOnList(arr -> ExpenseCalculator.sumArrayList(arr));
             return "You spent a total of $" + total;
         } else if (word.startsWith("spentOn")) {
             return handleSpentOn(word);
@@ -97,7 +97,7 @@ public class Interact {
      */
     public String handleSpentOn(String word) throws DukeException {
         LocalDate ld = Decoder.parseLD(word);
-        int total = expenseManager.operateOnList(arr -> ExpenseCalculator.spentDay(arr, ld));
+        double total = expenseManager.operateOnList(arr -> ExpenseCalculator.spentDay(arr, ld));
         return "You spent a total of $" + total;
     }
